refactor: Show only tick option for now
